✨ (client): responsive-header에서 sticky 속성 제거
- home-header에서 sticky 속성 따로 적용
